# Use Vectorized to Optimize multi input concat kernel