<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
Performance optimization

### PR changes
OPs

### Describe
使用向量化操作优化多个输入的Concat Kernel

使用IndexType Dispatch不同类型的索引，以节省寄存器

int64下用的寄存器40，occupancy被限制到了50%
int32下用的寄存器19，occupancy不受限制

TestCase来自TCIR的一段，等价代码为：
```python
x1 = paddle.ones((1, 90, 518400), dtype=paddle.float16)
x2 = paddle.ones((1, 32, 518400), dtype=paddle.float16)
x3 = paddle.ones((1, 32, 518400), dtype=paddle.float16)
x4 = paddle.ones((1, 32, 518400), dtype=paddle.float16)
x5 = paddle.ones((1, 32, 518400), dtype=paddle.float16)
out = paddle.concat([x1, x2, x3, x4, x5], axis=1)
```
|Paddle Before(us) |Paddle After(us)|OneFlow(us)|PyTorch(us)|
| --- | --- | --- | --- |
| 645.28us|  320us | 124+41+41+41+41=288us | 519us |